requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
SQLAlchemy
requests
audioop-lts
icalendar
selenium
cachetools